Move unflag-unbilled service
