+++ test/services/bill-runs/supplementary/process-bill-run.service.test.js
@@ -20,12 +20,12 @@ const FetchChargeVersionsService = require('../../../../app/services/bill-runs/s
 const HandleErroredBillRunService = require('../../../../app/services/bill-runs/handle-errored-bill-run.service.js')
 const LegacyRefreshBillRunRequest = require('../../../../app/requests/legacy/refresh-bill-run.request.js')
 const ProcessBillingPeriodService = require('../../../../app/services/bill-runs/supplementary/process-billing-period.service.js')
-const UnflagUnbilledLicencesService = require('../../../../app/services/bill-runs/supplementary/unflag-unbilled-licences.service.js')
+const UnflagUnbilledSupplementaryLicencesService = require('../../../../app/services/bill-runs/unflag-unbilled-supplementary-licences.service.js')
 
 // Thing under test
-const SupplementaryProcessBillRunService = require('../../../../app/services/bill-runs/supplementary/process-bill-run.service.js')
+const ProcessBillRunService = require('../../../../app/services/bill-runs/supplementary/process-bill-run.service.js')
 
-describe('Supplementary Process Bill Run service', () => {
+describe('Bill Runs - Supplementary - Process Bill Run service', () => {
   const billingPeriods = [
     { startDate: new Date('2023-04-01'), endDate: new Date('2024-03-31') },
     { startDate: new Date('2022-04-01'), endDate: new Date('2023-03-31') }
@@ -62,7 +62,7 @@ describe('Supplementary Process Bill Run service', () => {
   describe('when the service is called', () => {
     beforeEach(() => {
       Sinon.stub(FetchChargeVersionsService, 'go').resolves({ chargeVersions: [], licenceIdsForPeriod: [] })
-      Sinon.stub(UnflagUnbilledLicencesService, 'go')
+      Sinon.stub(UnflagUnbilledSupplementaryLicencesService, 'go')
     })
 
     describe('and nothing is billed', () => {
@@ -71,7 +71,7 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('sets the Bill Run status to empty', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const result = await BillRunModel.query().findById(billRun.id)
 
@@ -85,7 +85,7 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('sets the Bill Run status to processing', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const result = await BillRunModel.query().findById(billRun.id)
 
@@ -93,19 +93,19 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('tells the charging module API to "generate" the bill run', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         expect(chargingModuleGenerateBillRunRequestStub.called).to.be.true()
       })
 
       it('tells the legacy service to start its refresh job', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         expect(legacyRefreshBillRunRequestStub.called).to.be.true()
       })
 
       it('it logs the time taken', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const args = notifierStub.omg.firstCall.args
 
@@ -127,7 +127,7 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('calls HandleErroredBillRunService with appropriate error code', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const handlerArgs = handleErroredBillRunStub.firstCall.args
 
@@ -135,7 +135,7 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('logs the error', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const args = notifierStub.omfg.firstCall.args
 
@@ -158,7 +158,7 @@ describe('Supplementary Process Bill Run service', () => {
         })
 
         it('calls HandleErroredBillRunService with the error code', async () => {
-          await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+          await ProcessBillRunService.go(billRun, billingPeriods)
 
           const handlerArgs = handleErroredBillRunStub.firstCall.args
 
@@ -166,7 +166,7 @@ describe('Supplementary Process Bill Run service', () => {
         })
 
         it('logs the error', async () => {
-          await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+          await ProcessBillRunService.go(billRun, billingPeriods)
 
           const args = notifierStub.omfg.firstCall.args
 
@@ -186,11 +186,11 @@ describe('Supplementary Process Bill Run service', () => {
 
         Sinon.stub(FetchChargeVersionsService, 'go').resolves({ chargeVersions: [], licenceIdsForPeriod: [] })
         Sinon.stub(ProcessBillingPeriodService, 'go').resolves(false)
-        Sinon.stub(UnflagUnbilledLicencesService, 'go').rejects(thrownError)
+        Sinon.stub(UnflagUnbilledSupplementaryLicencesService, 'go').rejects(thrownError)
       })
 
       it('calls HandleErroredBillRunService without an error code', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const handlerArgs = handleErroredBillRunStub.firstCall.args
 
@@ -198,7 +198,7 @@ describe('Supplementary Process Bill Run service', () => {
       })
 
       it('logs the error', async () => {
-        await SupplementaryProcessBillRunService.go(billRun, billingPeriods)
+        await ProcessBillRunService.go(billRun, billingPeriods)
 
         const args = notifierStub.omfg.firstCall.args
